・クエリのテーブルからクエリを取得し、ループさせるように実装 ・各テーブルへの参照、登録処理、及び全体の動作確認完了 ・tweet_operation.py をクラス化 ・一部関数を他クラスに統合

+++ src/sql/dml/create_tweet_info.sql
@@ -1,5 +1,5 @@
 create table TWEET_INFO (
-    tweet_id integer Primary Key,
+    tweet_id bigint Primary Key,
     tweet_url varchar(100) Not Null,
     text varchar(280) Not Null,
     media_url varchar(60),
